memory problem addressed, cache of glm fit implemented and contrast computation refactored
